bug: ChatRoomBase spicy 자료형 float 수정

+++ app/schemas/chat.py
@@ -9,7 +9,7 @@ class ChatRoomBase(BaseModel):
     character_id: int
     character_name: str
     topic: Optional[str] = None
-    spicy: Optional[int] = None
+    spicy: Optional[float] = None
     created_at: datetime
     name: str
